Implementation of undistort video and video.reset
